+++ NEWS.md
@@ -1,5 +1,6 @@
 # fgeo.base 0.0.0.9001
 
+* `pick_*()` and `drop_*()` now work with both census and viewfull tables.
 * Functions to filter dataframes are now named as `pick_*()`.
 
 # fgeo.base 0.0.0.9000